+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/files/Sticker.kt
@@ -21,6 +21,8 @@ data class Sticker(
     val emoji: String? = null,
     @SerialName(stickerSetNameField)
     val stickerSetName: String? = null,
+    @SerialName(isAnimatedField)
+    val isAnimated: Boolean = false,
     @SerialName(maskPositionField)
     val maskPosition: MaskPosition? = null,
     @SerialName(fileSizeField)